Fix PriceList save handling in Django admin
Handle the products in the save_model method in the
PriceListAdmin class and change the PriceListForm save
so that it only handles the products when the changes
are committed.

Refs TTV-14

+++ respa_pricing/forms.py
@@ -46,26 +46,25 @@ class PriceListForm(forms.ModelForm):
         Create or archive the product (and handle the priced product)
         based on the selected resource.
         """
-        saved_form = super().save(commit=False)
+        saved_form = super().save(commit=commit)
         current_resource = None
         selected_resource = self.cleaned_data["resource"]
 
         if self.instance.pk and hasattr(self.instance, "priced_product"):
             current_resource = self.instance.priced_product.product.resources.first()
 
-        saved_form.save()
+        if commit:
+            if current_resource and selected_resource != current_resource:
+                # Archive existing products linked to the price list
+                self.instance.priced_product.product.archived_at = now()
+                self.instance.priced_product.product.save()
+                self.instance.priced_product.delete()
 
-        if current_resource and selected_resource != current_resource:
-            # Archive existing products linked to the price list
-            self.instance.priced_product.product.archived_at = now()
-            self.instance.priced_product.product.save()
-            self.instance.priced_product.delete()
-
-        if selected_resource and selected_resource != current_resource:
-            # Create new product
-            prod = Product.objects.create(name=selected_resource.name)
-            prod.resources.add(selected_resource)
-            PricedProduct.objects.create(product=prod, price_list=self.instance)
+            if selected_resource and selected_resource != current_resource:
+                # Create new product
+                prod = Product.objects.create(name=selected_resource.name)
+                prod.resources.add(selected_resource)
+                PricedProduct.objects.create(product=prod, price_list=self.instance)
 
         return saved_form
